# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Plantar Futuro API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://plantarfuturo.com/ws"
        },
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://swagger.local"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Login de usuarios",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::login",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Correo del usuario",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "401": {
                        "description": "Intento de login rechazado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Mostrar usuarios",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::index",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los usuarios."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Registrar usuario",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::store",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "document": {
                                        "description": "Documento del usuario",
                                        "type": "string",
                                        "format": "string",
                                        "example": "ASD123"
                                    },
                                    "name": {
                                        "description": "Nombre del usuario",
                                        "type": "string",
                                        "format": "string",
                                        "example": "CA"
                                    },
                                    "lastname": {
                                        "description": "Apellido del usuario",
                                        "type": "string",
                                        "format": "string",
                                        "example": "MC"
                                    },
                                    "phone": {
                                        "description": "Telefono del usuario",
                                        "type": "string",
                                        "format": "string",
                                        "example": "3145586963"
                                    },
                                    "address": {
                                        "description": "Direccion del usuario",
                                        "type": "string",
                                        "format": "string",
                                        "example": "Cra 1 Call 12"
                                    },
                                    "position_id": {
                                        "description": "Cargo del usuario",
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    },
                                    "email": {
                                        "description": "Email del usuario",
                                        "type": "string",
                                        "format": "string",
                                        "example": "camc@gmail.com"
                                    },
                                    "password": {
                                        "description": "Password del usuario",
                                        "type": "string",
                                        "format": "string",
                                        "example": "123456"
                                    },
                                    "document_type_id": {
                                        "description": "Tipo de documento",
                                        "type": "string",
                                        "format": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proyecto registrada."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/users/risk-assignation": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Asignacion de usuarios a proyectos de riesgos",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::riskAssignation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project_id": {
                                        "description": "Id del proyecto",
                                        "type": "integer"
                                    },
                                    "users": {
                                        "description": "Users ID's",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "items": {}
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "project_id": 1,
                                    "users": [
                                        1,
                                        2
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente registrada."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/users/assignation": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Asignacion de usuarios a proyectos",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::assignation",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project_id": {
                                        "description": "Id del proyecto",
                                        "type": "integer"
                                    },
                                    "users": {
                                        "description": "Users ID's",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "items": {}
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "project_id": 1,
                                    "users": [
                                        1,
                                        2
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente registrada."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/users/unassign": {
            "post": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Desasignacion de usuarios de proyectos",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::unassign",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "project_id": {
                                        "description": "Id del proyecto",
                                        "type": "integer"
                                    },
                                    "users": {
                                        "description": "Users ID's",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "items": {}
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "project_id": 1,
                                    "users": [
                                        1,
                                        2
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente registrada."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/users/projects/{id}": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Listado de proyectos por usuario",
                "operationId": "App\\Http\\Controllers\\Api\\AuthController::getProjects",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar todos los proyectos."
                    },
                    "204": {
                        "description": "No hay resultados que mostrar."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}